Factor out common plot operations to utils
