+++ crates/re_space_view_time_series/src/legacy_visualizer_system.rs
@@ -1,5 +1,3 @@
-use re_data_store::TimeRange;
-use re_log_types::{EntityPath, TimeInt};
 use re_query_cache::{MaybeCachedComponentData, QueryError};
 use re_types::{
     archetypes::TimeSeriesScalar,
@@ -7,14 +5,13 @@ use re_types::{
     Loggable,
 };
 use re_viewer_context::{
-    external::re_entity_db::TimeSeriesAggregator, AnnotationMap, DefaultColor,
-    IdentifiedViewSystem, ResolvedAnnotationInfo, SpaceViewSystemExecutionError, ViewQuery,
+    AnnotationMap, DefaultColor, IdentifiedViewSystem, SpaceViewSystemExecutionError, ViewQuery,
     ViewerContext, VisualizerQueryInfo, VisualizerSystem,
 };
 
 use crate::{
-    aggregation::{AverageAggregator, MinMaxAggregator},
-    overrides::lookup_override,
+    overrides::{initial_override_color, lookup_override},
+    util::{determine_time_range, points_to_lines},
     PlotPoint, PlotPointAttrs, PlotSeries, PlotSeriesKind,
 };
 
@@ -71,15 +68,7 @@ impl VisualizerSystem for LegacyTimeSeriesSystem {
         component: &re_types::ComponentName,
     ) -> Option<re_log_types::DataCell> {
         if *component == Color::name() {
-            let default_color = DefaultColor::EntityPath(entity_path);
-
-            let annotation_info = ResolvedAnnotationInfo::default();
-
-            let color = annotation_info.color(None, default_color);
-
-            let [r, g, b, a] = color.to_array();
-
-            Some([Color::from_unmultiplied_rgba(r, g, b, a)].into())
+            Some([initial_override_color(entity_path)].into())
         } else {
             None
         }
@@ -111,40 +100,7 @@ impl LegacyTimeSeriesSystem {
         for data_result in query.iter_visible_data_results(Self::identifier()) {
             let mut points = Vec::new();
 
-            let visible_history = match query.timeline.typ() {
-                re_log_types::TimeType::Time => {
-                    data_result.accumulated_properties().visible_history.nanos
-                }
-                re_log_types::TimeType::Sequence => {
-                    data_result
-                        .accumulated_properties()
-                        .visible_history
-                        .sequences
-                }
-            };
-
-            let mut time_range = if data_result.accumulated_properties().visible_history.enabled {
-                visible_history.time_range(query.latest_at)
-            } else {
-                TimeRange::new(TimeInt::MIN, TimeInt::MAX)
-            };
-
-            // TODO(cmc): We would love to reduce the query to match the actual plot bounds, but because
-            // the plot widget handles zoom after we provide it with data for the current frame,
-            // this results in an extremely jarring frame delay.
-            // Just try it out and you'll see what I mean.
-            if false {
-                if let Some(plot_bounds) = plot_bounds {
-                    time_range.min = TimeInt::max(
-                        time_range.min,
-                        (plot_bounds.range_x().start().floor() as i64).into(),
-                    );
-                    time_range.max = TimeInt::min(
-                        time_range.max,
-                        (plot_bounds.range_x().end().ceil() as i64).into(),
-                    );
-                }
-            }
+            let time_range = determine_time_range(query, data_result, plot_bounds);
 
             {
                 re_tracing::profile_scope!("primary", &data_result.entity_path.to_string());
@@ -244,182 +200,17 @@ impl LegacyTimeSeriesSystem {
                 )?;
             }
 
-            if points.is_empty() {
-                continue;
-            }
-
-            let (aggregator, aggregation_factor, points) = {
-                let aggregator = data_result
-                    .accumulated_properties()
-                    .time_series_aggregator
-                    .get();
-
-                let aggregation_factor = plot_value_delta;
-                let num_points_before = points.len() as f64;
-
-                let points = if aggregation_factor > 2.0 {
-                    re_tracing::profile_scope!("aggregate", aggregator.to_string());
-
-                    #[allow(clippy::match_same_arms)] // readability
-                    match aggregator {
-                        TimeSeriesAggregator::Off => points,
-                        TimeSeriesAggregator::Average => {
-                            AverageAggregator::aggregate(aggregation_factor, &points)
-                        }
-                        TimeSeriesAggregator::Min => {
-                            MinMaxAggregator::Min.aggregate(aggregation_factor, &points)
-                        }
-                        TimeSeriesAggregator::Max => {
-                            MinMaxAggregator::Max.aggregate(aggregation_factor, &points)
-                        }
-                        TimeSeriesAggregator::MinMax => {
-                            MinMaxAggregator::MinMax.aggregate(aggregation_factor, &points)
-                        }
-                        TimeSeriesAggregator::MinMaxAverage => {
-                            MinMaxAggregator::MinMaxAverage.aggregate(aggregation_factor, &points)
-                        }
-                    }
-                } else {
-                    points
-                };
-
-                let num_points_after = points.len() as f64;
-                let actual_aggregation_factor = num_points_before / num_points_after;
-
-                re_log::trace!(
-                    id = %query.space_view_id,
-                    plot_value_delta,
-                    ?aggregator,
-                    aggregation_factor,
-                    num_points_before,
-                    num_points_after,
-                    actual_aggregation_factor,
-                );
-
-                (aggregator, actual_aggregation_factor, points)
-            };
-
-            re_tracing::profile_scope!("secondary", &data_result.entity_path.to_string());
-
-            let min_time = store
-                .entity_min_time(&query.timeline, &data_result.entity_path)
-                .map_or(points.first().map_or(0, |p| p.time), |time| time.as_i64());
-
-            // If all points within a line share the label (and it isn't `None`), then we use it
-            // as the whole line label for the plot legend.
-            // Otherwise, we just use the entity path as-is.
-            let same_label = |points: &[PlotPoint]| -> Option<String> {
-                let label = points[0].attrs.label.as_ref()?;
-                (points.iter().all(|p| p.attrs.label.as_ref() == Some(label)))
-                    .then(|| label.clone())
-            };
-            let line_label =
-                same_label(&points).unwrap_or_else(|| data_result.entity_path.to_string());
-
-            if points.len() == 1 {
-                self.lines.push(PlotSeries {
-                    label: line_label,
-                    color: points[0].attrs.color,
-                    width: 2.0 * points[0].attrs.radius,
-                    kind: PlotSeriesKind::Scatter,
-                    points: vec![(points[0].time, points[0].value)],
-                    entity_path: data_result.entity_path.clone(),
-                    aggregator: *aggregator,
-                    aggregation_factor,
-                    min_time,
-                });
-            } else {
-                self.add_line_segments(
-                    &line_label,
-                    points,
-                    &data_result.entity_path,
-                    *aggregator,
-                    aggregation_factor,
-                    min_time,
-                );
-            }
+            // Now convert the `PlotPoints` into lines
+            points_to_lines(
+                data_result,
+                plot_value_delta,
+                points,
+                store,
+                query,
+                &mut self.lines,
+            );
         }
 
         Ok(())
     }
-
-    // We have a bunch of raw points, and now we need to group them into actual line
-    // segments.
-    // A line segment is a continuous run of points with identical attributes: each time
-    // we notice a change in attributes, we need a new line segment.
-    #[inline(never)] // Better callstacks on crashes
-    fn add_line_segments(
-        &mut self,
-        line_label: &str,
-        points: Vec<PlotPoint>,
-        entity_path: &EntityPath,
-        aggregator: TimeSeriesAggregator,
-        aggregation_factor: f64,
-        min_time: i64,
-    ) {
-        re_tracing::profile_function!();
-
-        let num_points = points.len();
-        let mut attrs = points[0].attrs.clone();
-        let mut line: PlotSeries = PlotSeries {
-            label: line_label.to_owned(),
-            color: attrs.color,
-            width: 2.0 * attrs.radius,
-            points: Vec::with_capacity(num_points),
-            kind: attrs.kind,
-            entity_path: entity_path.clone(),
-            aggregator,
-            aggregation_factor,
-            min_time,
-        };
-
-        for (i, p) in points.into_iter().enumerate() {
-            if p.attrs == attrs {
-                // Same attributes, just add to the current line segment.
-
-                line.points.push((p.time, p.value));
-            } else {
-                // Attributes changed since last point, break up the current run into a
-                // line segment, and start the next one.
-
-                attrs = p.attrs;
-                let prev_line = std::mem::replace(
-                    &mut line,
-                    PlotSeries {
-                        label: line_label.to_owned(),
-                        color: attrs.color,
-                        width: 2.0 * attrs.radius,
-                        kind: attrs.kind,
-                        points: Vec::with_capacity(num_points - i),
-                        entity_path: entity_path.clone(),
-                        aggregator,
-                        aggregation_factor,
-                        min_time,
-                    },
-                );
-
-                let cur_continuous = matches!(attrs.kind, PlotSeriesKind::Continuous);
-                let prev_continuous = matches!(prev_line.kind, PlotSeriesKind::Continuous);
-
-                let prev_point = *prev_line.points.last().unwrap();
-                self.lines.push(prev_line);
-
-                // If the previous point was continuous and the current point is continuous
-                // too, then we want the 2 segments to appear continuous even though they
-                // are actually split from a data standpoint.
-                if cur_continuous && prev_continuous {
-                    line.points.push(prev_point);
-                }
-
-                // Add the point that triggered the split to the new segment.
-                line.points.push((p.time, p.value));
-            }
-        }
-
-        if !line.points.is_empty() {
-            self.lines.push(line);
-        }
-    }
 }
-
-// ---
